refactor: Separate details of http from adapter
Separate some implementation details of node-http-adapter
from adapter class to HttpWrapper and TransportWrapper classes
